Thumbnails are degrade
Looks like there's a weird behavior of node-gm when used async. It takes the last created thumb to create the next. The result is a pixelized images when the smallest one is... really small.

To overcome this issue, I've just reloaded the original file before applying the gm(...).thumb(...) function.